Move upload manager to its own package (#1732)
This PR moves the upload manager to our `internal` package.

Related https://github.com/SiaFoundation/renterd/issues/1646

---------
