+++ .changeset/move_upload_manager_to_internal_package.md
@@ -1,0 +1,5 @@
+---
+default: patch
+---
+
+# Move upload manager to internal package
