+++ internal/upload/mimereader.go
@@ -1,0 +1,15 @@
+package upload
+
+import (
+	"bytes"
+	"io"
+
+	"github.com/gabriel-vasile/mimetype"
+)
+
+func NewMimeReader(r io.Reader) (mimeType string, recycled io.Reader, err error) {
+	buf := bytes.NewBuffer(nil)
+	mtype, err := mimetype.DetectReader(io.TeeReader(r, buf))
+	recycled = io.MultiReader(buf, r)
+	return mtype.String(), recycled, err
+}
