+++ worker/bench_test.go
@@ -24,20 +24,20 @@ func (z *zeroReader) Read(p []byte) (n int, err error) {
 // object.
 // 1036.74 MB/s | M2 Pro | c9dc1b6
 func BenchmarkDownloaderSingleObject(b *testing.B) {
-	w := newTestWorker(b)
+	w := newTestWorker(b, newTestWorkerCfg())
 
 	up := testParameters(b.TempDir())
-	up.rs.MinShards = 10
-	up.rs.TotalShards = 30
-	up.packing = false
-	w.AddHosts(up.rs.TotalShards)
+	up.RS.MinShards = 10
+	up.RS.TotalShards = 30
+	up.Packing = false
+	w.AddHosts(up.RS.TotalShards)
 
-	data := bytes.NewReader(frand.Bytes(int(up.rs.SlabSizeNoRedundancy())))
-	_, _, err := w.uploadManager.Upload(context.Background(), data, w.Contracts(), up)
+	data := bytes.NewReader(frand.Bytes(int(up.RS.SlabSizeNoRedundancy())))
+	_, _, err := w.uploadManager.Upload(context.Background(), data, w.UploadHosts(), up)
 	if err != nil {
 		b.Fatal(err)
 	}
-	o, err := w.os.Object(context.Background(), testBucket, up.key, api.GetObjectOptions{})
+	o, err := w.os.Object(context.Background(), testBucket, up.Key, api.GetObjectOptions{})
 	if err != nil {
 		b.Fatal(err)
 	}
@@ -57,19 +57,19 @@ func BenchmarkDownloaderSingleObject(b *testing.B) {
 // Speed       | CPU    | Commit
 // 433.86 MB/s | M2 Pro | bae6e77
 func BenchmarkUploaderSingleObject(b *testing.B) {
-	w := newTestWorker(b)
+	w := newTestWorker(b, newTestWorkerCfg())
 
 	up := testParameters(b.TempDir())
-	up.rs.MinShards = 10
-	up.rs.TotalShards = 30
-	up.packing = false
-	w.AddHosts(up.rs.TotalShards)
+	up.RS.MinShards = 10
+	up.RS.TotalShards = 30
+	up.Packing = false
+	w.AddHosts(up.RS.TotalShards)
 
-	data := io.LimitReader(&zeroReader{}, int64(b.N*rhpv2.SectorSize*up.rs.MinShards))
-	b.SetBytes(int64(rhpv2.SectorSize * up.rs.MinShards))
+	data := io.LimitReader(&zeroReader{}, int64(b.N*rhpv2.SectorSize*up.RS.MinShards))
+	b.SetBytes(int64(rhpv2.SectorSize * up.RS.MinShards))
 	b.ResetTimer()
 
-	_, _, err := w.uploadManager.Upload(context.Background(), data, w.Contracts(), up)
+	_, _, err := w.uploadManager.Upload(context.Background(), data, w.UploadHosts(), up)
 	if err != nil {
 		b.Fatal(err)
 	}
@@ -80,20 +80,20 @@ func BenchmarkUploaderSingleObject(b *testing.B) {
 // Speed       | CPU    | Commit
 // 282.47 MB/s | M2 Pro | bae6e77
 func BenchmarkUploaderMultiObject(b *testing.B) {
-	w := newTestWorker(b)
+	w := newTestWorker(b, newTestWorkerCfg())
 
 	up := testParameters(b.TempDir())
-	up.rs.MinShards = 10
-	up.rs.TotalShards = 30
-	up.packing = false
-	w.AddHosts(up.rs.TotalShards)
+	up.RS.MinShards = 10
+	up.RS.TotalShards = 30
+	up.Packing = false
+	w.AddHosts(up.RS.TotalShards)
 
-	b.SetBytes(int64(rhpv2.SectorSize * up.rs.MinShards))
+	b.SetBytes(int64(rhpv2.SectorSize * up.RS.MinShards))
 	b.ResetTimer()
 
 	for i := 0; i < b.N; i++ {
-		data := io.LimitReader(&zeroReader{}, int64(rhpv2.SectorSize*up.rs.MinShards))
-		_, _, err := w.uploadManager.Upload(context.Background(), data, w.Contracts(), up)
+		data := io.LimitReader(&zeroReader{}, int64(rhpv2.SectorSize*up.RS.MinShards))
+		_, _, err := w.uploadManager.Upload(context.Background(), data, w.UploadHosts(), up)
 		if err != nil {
 			b.Fatal(err)
 		}
