--- worker/upload_params.go
@@ -1,88 +1,0 @@
-package worker
-
-import (
-	"go.sia.tech/renterd/api"
-	"go.sia.tech/renterd/object"
-)
-
-type uploadParameters struct {
-	bucket string
-	key    string
-
-	multipart  bool
-	uploadID   string
-	partNumber int
-
-	ec               object.EncryptionKey
-	encryptionOffset uint64
-
-	rs       api.RedundancySettings
-	bh       uint64
-	packing  bool
-	mimeType string
-
-	metadata api.ObjectUserMetadata
-}
-
-func defaultParameters(bucket, key string, rs api.RedundancySettings) uploadParameters {
-	return uploadParameters{
-		bucket: bucket,
-		key:    key,
-
-		ec:               object.GenerateEncryptionKey(object.EncryptionKeyTypeSalted), // random key
-		encryptionOffset: 0,                                                            // from the beginning
-
-		rs: rs,
-	}
-}
-
-type UploadOption func(*uploadParameters)
-
-func WithBlockHeight(bh uint64) UploadOption {
-	return func(up *uploadParameters) {
-		up.bh = bh
-	}
-}
-
-func WithCustomKey(ec object.EncryptionKey) UploadOption {
-	return func(up *uploadParameters) {
-		up.ec = ec
-	}
-}
-
-func WithCustomEncryptionOffset(offset uint64) UploadOption {
-	return func(up *uploadParameters) {
-		up.encryptionOffset = offset
-	}
-}
-
-func WithMimeType(mimeType string) UploadOption {
-	return func(up *uploadParameters) {
-		up.mimeType = mimeType
-	}
-}
-
-func WithPacking(packing bool) UploadOption {
-	return func(up *uploadParameters) {
-		up.packing = packing
-	}
-}
-
-func WithPartNumber(partNumber int) UploadOption {
-	return func(up *uploadParameters) {
-		up.partNumber = partNumber
-	}
-}
-
-func WithUploadID(uploadID string) UploadOption {
-	return func(up *uploadParameters) {
-		up.uploadID = uploadID
-		up.multipart = true
-	}
-}
-
-func WithObjectUserMetadata(metadata api.ObjectUserMetadata) UploadOption {
-	return func(up *uploadParameters) {
-		up.metadata = metadata
-	}
-}
